+++ src/Services/ServiceBuilderFactory.php
@@ -25,6 +25,7 @@ use Bitrix24\SDK\Core\Exceptions\InvalidArgumentException;
 use Psr\Log\LoggerInterface;
 use Psr\Log\NullLogger;
 use Symfony\Component\EventDispatcher\EventDispatcher;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
 
 class ServiceBuilderFactory
@@ -66,7 +67,7 @@ class ServiceBuilderFactory
      * Init service builder from request
      *
      * @param ApplicationProfile $applicationProfile
-     * @param AuthToken $accessToken
+     * @param AuthToken $authToken
      * @param string $bitrix24DomainUrl
      *
      * @return ServiceBuilder
@@ -74,13 +75,13 @@ class ServiceBuilderFactory
      */
     public function initFromRequest(
         ApplicationProfile $applicationProfile,
-        AuthToken          $accessToken,
+        AuthToken          $authToken,
         string             $bitrix24DomainUrl
     ): ServiceBuilder
     {
         return $this->getServiceBuilder(
             Credentials::createFromOAuth(
-                $accessToken,
+                $authToken,
                 $applicationProfile,
                 $bitrix24DomainUrl
             )
@@ -135,7 +136,10 @@ class ServiceBuilderFactory
      * @param LoggerInterface|null $logger optional logger for debug logs
      * @throws InvalidArgumentException
      */
-    public static function createServiceBuilderFromWebhook(string $webhookUrl, ?EventDispatcherInterface $eventDispatcher = null, ?LoggerInterface $logger = null): ServiceBuilder
+    public static function createServiceBuilderFromWebhook(
+        string                    $webhookUrl,
+        ?EventDispatcherInterface $eventDispatcher = null,
+        ?LoggerInterface          $logger = null): ServiceBuilder
     {
         if ($eventDispatcher === null) {
             $eventDispatcher = new EventDispatcher();
@@ -144,5 +148,46 @@ class ServiceBuilderFactory
             $logger = new NullLogger();
         }
         return (new ServiceBuilderFactory($eventDispatcher, $logger))->initFromWebhook($webhookUrl);
+    }
+
+    /**
+     * Create service builder from placement request
+     *
+     * @param Request $placementRequest The placement request object that contains the request data.
+     * @param ApplicationProfile $applicationProfile The application profile object.
+     * @param EventDispatcherInterface|null $eventDispatcher Optional event dispatcher for subscribing to domain events.
+     * @param LoggerInterface|null $logger Optional logger for debug logs.
+     * @return ServiceBuilder The service builder object.
+     * @throws InvalidArgumentException If the key "DOMAIN" is not found in the request.
+     */
+    public static function createServiceBuilderFromPlacementRequest(
+        Request                   $placementRequest,
+        ApplicationProfile        $applicationProfile,
+        ?EventDispatcherInterface $eventDispatcher = null,
+        ?LoggerInterface          $logger = null
+    ): ServiceBuilder
+    {
+        if (!array_key_exists('DOMAIN', $placementRequest->query->keys())) {
+            throw new InvalidArgumentException('key «DOMAIN» not found in request');
+        }
+
+        $rawDomainUrl = trim((string)$placementRequest->query->get('DOMAIN'));
+        if ($rawDomainUrl === '') {
+            throw new InvalidArgumentException('DOMAIN key cannot be empty in request');
+        }
+
+        if ($eventDispatcher === null) {
+            $eventDispatcher = new EventDispatcher();
+        }
+        if ($logger === null) {
+            $logger = new NullLogger();
+        }
+
+        return (new ServiceBuilderFactory($eventDispatcher, $logger))
+            ->initFromRequest(
+                $applicationProfile,
+                AuthToken::initFromPlacementRequest($placementRequest),
+                $rawDomainUrl
+            );
     }
 }